Tweak display of status messages, add a "finished" state
The biggest change is the addition of a "finished" state, which is simply a torrent that has finished downloading and is paused. As a result, there is also a new hotkey, Cmd/Ctrl+Alt+Backspace, which selects all the finished torrents and deletes them. The progress bar now also reflects seeding state (if you have set a seeding ratio), among other things.